add fs_to_narrow() to convert multibyte to ASCII
Windows: WideCharToMultiByte to convert PWSTR to std::string

non-Windows: use deprecated std::wstring_convert for now

+++ src/partition.cpp
@@ -90,7 +90,7 @@ std::string fs_filesystem_type(std::string_view path)
 
   std::string name(MAX_PATH+1, '\0');
   // https://learn.microsoft.com/en-us/windows/win32/api/fileapi/nf-fileapi-getvolumeinformationa
-  if(GetVolumeInformationA(r.data(), nullptr, 0, nullptr, nullptr, nullptr, name.data(), name.size()))
+  if(GetVolumeInformationA(r.data(), nullptr, 0, nullptr, nullptr, nullptr, name.data(), static_cast<DWORD>(name.size())))
     return fs_trim(name);
 
   fs_print_error(path, "filesystem_type");
